Move command handler to the package level.
